+++ src/main/java/com/tsp/new_tsp_admin/api/faq/AdminFaqJpaController.java
@@ -10,10 +10,12 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.client.HttpClientErrorException;
 
 import javax.validation.Valid;
+import java.net.URI;
 import java.rmi.ServerError;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -33,11 +35,11 @@ public class AdminFaqJpaController {
 
     /**
      * <pre>
-     * 1. MethodName : findFaqsList
+     * 1. MethodName : findFaqList
      * 2. ClassName  : AdminFaqJpaController.java
      * 3. Comment    : 관리자 FAQ 리스트 조회
-     * 4. 작성자       : CHO
-     * 5. 작성일       : 2022. 08. 22.
+     * 4. 작성자      : CHO
+     * 5. 작성일      : 2022. 08. 22.
      * </pre>
      */
     @ApiOperation(value = "FAQ 조회", notes = "FAQ를 조회한다.")
@@ -46,10 +48,11 @@ public class AdminFaqJpaController {
             @ApiResponse(code = 400, message = "잘못된 요청", response = HttpClientErrorException.BadRequest.class),
             @ApiResponse(code = 401, message = "허용되지 않는 관리자", response = HttpClientErrorException.Unauthorized.class),
             @ApiResponse(code = 403, message = "접근거부", response = HttpClientErrorException.class),
+            @ApiResponse(code = 404, message = "존재 하지 않음", response = HttpClientErrorException.NotFound.class),
             @ApiResponse(code = 500, message = "서버 에러", response = ServerError.class)
     })
     @GetMapping(value = "/lists")
-    public Map<String, Object> findFaqsList(@RequestParam(required = false) Map<String, Object> paramMap, Page page) {
+    public ResponseEntity<Map<String, Object>> findFaqList(@RequestParam(required = false) Map<String, Object> paramMap, Page page) {
         Map<String, Object> faqMap = new HashMap<>();
 
         int faqCount = this.adminFaqJpaService.findFaqCount(searchCommon.searchCommon(page, paramMap));
@@ -68,7 +71,7 @@ public class AdminFaqJpaController {
 
         faqMap.put("faqList", faqList);
 
-        return faqMap;
+        return ResponseEntity.ok().body(faqMap);
     }
 
     /**
@@ -76,21 +79,22 @@ public class AdminFaqJpaController {
      * 1. MethodName : findOneFaq
      * 2. ClassName  : AdminFaqJpaController.java
      * 3. Comment    : 관리자 FAQ 상세 조회
-     * 4. 작성자       : CHO
-     * 5. 작성일       : 2022. 08. 22.
+     * 4. 작성자      : CHO
+     * 5. 작성일      : 2022. 08. 22.
      * </pre>
      */
     @ApiOperation(value = "FAQ 상세 조회", notes = "FAQ를 상세 조회한다.")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "FAQ 상세 조회 성공", response = Map.class),
+            @ApiResponse(code = 200, message = "FAQ 상세 조회 성공", response = AdminFaqDTO.class),
             @ApiResponse(code = 400, message = "잘못된 요청", response = HttpClientErrorException.BadRequest.class),
             @ApiResponse(code = 401, message = "허용되지 않는 관리자", response = HttpClientErrorException.Unauthorized.class),
             @ApiResponse(code = 403, message = "접근거부", response = HttpClientErrorException.class),
+            @ApiResponse(code = 404, message = "존재 하지 않음", response = HttpClientErrorException.NotFound.class),
             @ApiResponse(code = 500, message = "서버 에러", response = ServerError.class)
     })
     @GetMapping("/{idx}")
-    public AdminFaqDTO findOneFaq(@PathVariable Long idx) {
-        return adminFaqJpaService.findOneFaq(idx);
+    public ResponseEntity<AdminFaqDTO> findOneFaq(@PathVariable Long idx) {
+        return ResponseEntity.ok(adminFaqJpaService.findOneFaq(idx));
     }
 
     /**
@@ -98,21 +102,22 @@ public class AdminFaqJpaController {
      * 1. MethodName : findPrevOneFaq
      * 2. ClassName  : AdminFaqJpaController.java
      * 3. Comment    : 관리자 이전 FAQ 상세 조회
-     * 4. 작성자       : CHO
-     * 5. 작성일       : 2022. 09. 18.
+     * 4. 작성자      : CHO
+     * 5. 작성일      : 2022. 09. 18.
      * </pre>
      */
     @ApiOperation(value = "이전 FAQ 상세 조회", notes = "이전 FAQ를 상세 조회한다.")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "이전 FAQ 상세 조회 성공", response = Map.class),
+            @ApiResponse(code = 200, message = "이전 FAQ 상세 조회 성공", response = AdminFaqDTO.class),
             @ApiResponse(code = 400, message = "잘못된 요청", response = HttpClientErrorException.BadRequest.class),
             @ApiResponse(code = 401, message = "허용되지 않는 관리자", response = HttpClientErrorException.Unauthorized.class),
             @ApiResponse(code = 403, message = "접근거부", response = HttpClientErrorException.class),
+            @ApiResponse(code = 404, message = "존재 하지 않음", response = HttpClientErrorException.NotFound.class),
             @ApiResponse(code = 500, message = "서버 에러", response = ServerError.class)
     })
     @GetMapping("/{idx}/prev")
-    public AdminFaqDTO findPrevOneFaq(@PathVariable Long idx) {
-        return adminFaqJpaService.findPrevOneFaq(idx);
+    public ResponseEntity<AdminFaqDTO> findPrevOneFaq(@PathVariable Long idx) {
+        return ResponseEntity.ok(adminFaqJpaService.findPrevOneFaq(idx));
     }
 
     /**
@@ -120,21 +125,22 @@ public class AdminFaqJpaController {
      * 1. MethodName : findNextOneFaq
      * 2. ClassName  : AdminFaqJpaController.java
      * 3. Comment    : 관리자 다음 FAQ 상세 조회
-     * 4. 작성자       : CHO
-     * 5. 작성일       : 2022. 09. 18.
+     * 4. 작성자      : CHO
+     * 5. 작성일      : 2022. 09. 18.
      * </pre>
      */
     @ApiOperation(value = "다음 FAQ 상세 조회", notes = "다음 FAQ를 상세 조회한다.")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "다음 FAQ 상세 조회 성공", response = Map.class),
+            @ApiResponse(code = 200, message = "다음 FAQ 상세 조회 성공", response = AdminFaqDTO.class),
             @ApiResponse(code = 400, message = "잘못된 요청", response = HttpClientErrorException.BadRequest.class),
             @ApiResponse(code = 401, message = "허용되지 않는 관리자", response = HttpClientErrorException.Unauthorized.class),
             @ApiResponse(code = 403, message = "접근거부", response = HttpClientErrorException.class),
+            @ApiResponse(code = 404, message = "존재 하지 않음", response = HttpClientErrorException.NotFound.class),
             @ApiResponse(code = 500, message = "서버 에러", response = ServerError.class)
     })
     @GetMapping("/{idx}/next")
-    public AdminFaqDTO findNextOneFaq(@PathVariable Long idx) {
-        return adminFaqJpaService.findNextOneFaq(idx);
+    public ResponseEntity<AdminFaqDTO> findNextOneFaq(@PathVariable Long idx) {
+        return ResponseEntity.ok(adminFaqJpaService.findNextOneFaq(idx));
     }
 
     /**
@@ -142,21 +148,22 @@ public class AdminFaqJpaController {
      * 1. MethodName : insertFaq
      * 2. ClassName  : AdminFaqJpaController.java
      * 3. Comment    : 관리자 FAQ 저장
-     * 4. 작성자       : CHO
-     * 5. 작성일       : 2022. 08. 22.
+     * 4. 작성자      : CHO
+     * 5. 작성일      : 2022. 08. 22.
      * </pre>
      */
     @ApiOperation(value = "FAQ 저장", notes = "FAQ를 저장한다.")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "FAQ 등록성공", response = Map.class),
+            @ApiResponse(code = 201, message = "FAQ 등록성공", response = AdminFaqDTO.class),
             @ApiResponse(code = 400, message = "잘못된 요청", response = HttpClientErrorException.BadRequest.class),
             @ApiResponse(code = 401, message = "허용되지 않는 관리자", response = HttpClientErrorException.Unauthorized.class),
             @ApiResponse(code = 403, message = "접근거부", response = HttpClientErrorException.class),
+            @ApiResponse(code = 404, message = "존재 하지 않음", response = HttpClientErrorException.NotFound.class),
             @ApiResponse(code = 500, message = "서버 에러", response = ServerError.class)
     })
     @PostMapping(consumes = APPLICATION_JSON_VALUE)
-    public AdminFaqDTO insertFaq(@Valid @RequestBody AdminFaqEntity adminFaqEntity) {
-        return this.adminFaqJpaService.insertFaq(adminFaqEntity);
+    public ResponseEntity<AdminFaqDTO> insertFaq(@Valid @RequestBody AdminFaqEntity adminFaqEntity) {
+        return ResponseEntity.created(URI.create("")).body(adminFaqJpaService.insertFaq(adminFaqEntity));
     }
 
     /**
@@ -164,21 +171,25 @@ public class AdminFaqJpaController {
      * 1. MethodName : updateFaq
      * 2. ClassName  : AdminFaqJpaController.java
      * 3. Comment    : 관리자 FAQ 수정
-     * 4. 작성자       : CHO
-     * 5. 작성일       : 2022. 08. 16.
+     * 4. 작성자      : CHO
+     * 5. 작성일      : 2022. 08. 16.
      * </pre>
      */
     @ApiOperation(value = "FAQ 수정", notes = "FAQ를 수정한다.")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "FAQ 수정 성공", response = Map.class),
+            @ApiResponse(code = 200, message = "FAQ 수정 성공", response = AdminFaqDTO.class),
             @ApiResponse(code = 400, message = "잘못된 요청", response = HttpClientErrorException.BadRequest.class),
             @ApiResponse(code = 401, message = "허용되지 않는 관리자", response = HttpClientErrorException.Unauthorized.class),
             @ApiResponse(code = 403, message = "접근거부", response = HttpClientErrorException.class),
+            @ApiResponse(code = 404, message = "존재 하지 않음", response = HttpClientErrorException.NotFound.class),
             @ApiResponse(code = 500, message = "서버 에러", response = ServerError.class)
     })
     @PutMapping(value = "/{idx}", consumes = APPLICATION_JSON_VALUE)
-    public AdminFaqDTO updateFaq(@Valid @RequestBody AdminFaqEntity adminFaqEntity) {
-        return adminFaqJpaService.updateFaq(adminFaqEntity);
+    public ResponseEntity<AdminFaqDTO> updateFaq(@PathVariable Long idx, @Valid @RequestBody AdminFaqEntity adminFaqEntity) {
+        if (adminFaqJpaService.findOneFaq(idx) == null) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(adminFaqJpaService.updateFaq(adminFaqEntity));
     }
 
     /**
@@ -186,20 +197,25 @@ public class AdminFaqJpaController {
      * 1. MethodName : deleteFaq
      * 2. ClassName  : AdminFaqJpaController.java
      * 3. Comment    : 관리자 FAQ 삭제
-     * 4. 작성자       : CHO
-     * 5. 작성일       : 2022. 08. 22.
+     * 4. 작성자      : CHO
+     * 5. 작성일      : 2022. 08. 22.
      * </pre>
      */
     @ApiOperation(value = "FAQ 삭제", notes = "FAQ를 삭제 한다.")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "FAQ 삭제 성공", response = Map.class),
+            @ApiResponse(code = 204, message = "FAQ 삭제 성공", response = Long.class),
             @ApiResponse(code = 400, message = "잘못된 요청", response = HttpClientErrorException.BadRequest.class),
             @ApiResponse(code = 401, message = "허용되지 않는 관리자", response = HttpClientErrorException.Unauthorized.class),
             @ApiResponse(code = 403, message = "접근거부", response = HttpClientErrorException.class),
+            @ApiResponse(code = 404, message = "존재 하지 않음", response = HttpClientErrorException.NotFound.class),
             @ApiResponse(code = 500, message = "서버 에러", response = ServerError.class)
     })
     @DeleteMapping(value = "/{idx}")
-    public Long deleteFaq(@PathVariable Long idx) {
-        return adminFaqJpaService.deleteFaq(idx);
+    public ResponseEntity<Long> deleteFaq(@PathVariable Long idx) {
+        if (adminFaqJpaService.findOneFaq(idx) == null) {
+            return ResponseEntity.notFound().build();
+        }
+        adminFaqJpaService.deleteFaq(idx);
+        return ResponseEntity.noContent().build();
     }
 }
